renamed Slider to SliderUnstyled
