+++ packages/material-ui-lab/src/SliderStyled/SliderStyled.js
@@ -8,7 +8,7 @@ import {
   darken,
 } from '@material-ui/core/styles';
 import { capitalize } from '@material-ui/core/utils';
-import SliderUnstyled from '@material-ui/unstyled/Slider';
+import SliderUnstyled from '@material-ui/unstyled/SliderUnstyled';
 import ValueLabelStyled from './ValueLabelStyled';
 
 const overridesResolver = (props, styles, name) => {
