--- packages/material-ui-unstyled/src/Slider/index.d.ts
@@ -1,4 +1,0 @@
-export { default } from './Slider';
-export * from './Slider';
-export { default as ValueLabel } from './ValueLabel';
-export * from './ValueLabel';
