--- packages/material-ui-unstyled/src/Slider/index.js
@@ -1,2 +1,0 @@
-export { default } from './Slider';
-export { default as ValueLabel } from './ValueLabel';
