+++ packages/material-ui-unstyled/src/SliderUnstyled/ValueLabelUnstyled.d.ts
