+++ packages/material-ui-unstyled/src/SliderUnstyled/ValueLabelUnstyled.js
