+++ packages/material-ui-unstyled/src/SliderUnstyled/index.d.ts
@@ -1,0 +1,4 @@
+export { default } from './Slider';
+export * from './Slider';
+export { default as ValueLabelUnstyled } from './ValueLabelUnstyled';
+export * from './ValueLabelUnstyled';
