+++ packages/material-ui-unstyled/src/SliderUnstyled/index.js
@@ -1,0 +1,2 @@
+export { default } from './SliderUnstyled';
+export { default as ValueLabelUnstyled } from './ValueLabelUnstyled';
