+++ packages/material-ui-unstyled/src/index.js
@@ -1,2 +1,2 @@
-export { default as Slider } from './Slider';
-export * from './Slider';
+export { default as SliderUnstyled } from './SliderUnstyled';
+export * from './SliderUnstyled';
